+++ instrumentation/extensions/src/main/kotlin/de/mannodermaus/junit5/extensions/GrantPermissionExtension.kt
@@ -3,7 +3,6 @@ package de.mannodermaus.junit5.extensions
 import android.Manifest
 import android.annotation.SuppressLint
 import android.os.Build
-import androidx.annotation.VisibleForTesting
 import androidx.test.annotation.ExperimentalTestApi
 import androidx.test.internal.platform.ServiceLoaderWrapper.loadSingleService
 import androidx.test.internal.platform.content.PermissionGranter